Addresses are sinks POC (based on flume changes)
