+++ src/address.rs
@@ -1,19 +1,20 @@
 //! An address to an actor is a way to send it a message. An address allows an actor to be sent any
 //! kind of message that it can receive.
 
+use flume::r#async::SendSink;
+use futures_core::Stream;
+use futures_sink::Sink;
+use futures_util::{future, StreamExt};
 use std::fmt::{self, Display, Formatter};
 use std::future::Future;
+use std::pin::Pin;
+use std::task::{Context, Poll};
 use std::{cmp::Ordering, error::Error, hash::Hash};
 
-use flume::Sender;
-use futures_core::Stream;
-use futures_util::{future, StreamExt};
-
-use crate::envelope::ReturningEnvelope;
+use crate::envelope::{NonReturningEnvelope, ReturningEnvelope};
 use crate::manager::AddressMessage;
 use crate::refcount::{Either, RefCounter, Strong, Weak};
 use crate::send_future::{NameableSending, ResolveToHandlerReturn, SendFuture};
-use crate::sink::AddressSink;
 use crate::{Handler, KeepRunning};
 
 /// The actor is no longer running and disconnected from the sending address. For why this could
@@ -39,8 +40,8 @@ impl Error for Disconnected {}
 /// should be used instead. An address is created by calling the
 /// [`Actor::create`](../trait.Actor.html#method.create) or
 /// [`Context::run`](../struct.Context.html#method.run) methods, or by cloning another `Address`.
-pub struct Address<A, Rc: RefCounter = Strong> {
-    pub(crate) sender: Sender<AddressMessage<A>>,
+pub struct Address<A: 'static, Rc: RefCounter = Strong> {
+    pub(crate) sink: SendSink<'static, AddressMessage<A>>,
     pub(crate) ref_counter: Rc,
 }
 
@@ -57,7 +58,7 @@ impl<A> Address<A, Strong> {
     /// addresses exist.
     pub fn downgrade(&self) -> WeakAddress<A> {
         WeakAddress {
-            sender: self.sender.clone(),
+            sink: self.sink.clone(),
             ref_counter: self.ref_counter.downgrade(),
         }
     }
@@ -68,7 +69,7 @@ impl<A> Address<A, Either> {
     /// Converts this address into a weak address.
     pub fn downgrade(&self) -> WeakAddress<A> {
         WeakAddress {
-            sender: self.sender.clone(),
+            sink: self.sink.clone(),
             ref_counter: self.ref_counter.clone().into_weak(),
         }
     }
@@ -104,17 +105,17 @@ impl<A, Rc: RefCounter> Address<A, Rc> {
     /// })
     /// ```
     pub fn is_connected(&self) -> bool {
-        !self.sender.is_disconnected()
+        !self.sink.is_disconnected()
     }
 
     /// Returns the number of messages in the actor's mailbox.
     pub fn len(&self) -> usize {
-        self.sender.len()
+        self.sink.len()
     }
 
     /// The total capacity of the actor's mailbox.
     pub fn capacity(&self) -> Option<usize> {
-        self.sender.capacity()
+        self.sink.capacity()
     }
 
     /// Returns whether the actor's mailbox is empty.
@@ -125,8 +126,8 @@ impl<A, Rc: RefCounter> Address<A, Rc> {
     /// Convert this address into a generic address which can be weak or strong.
     pub fn as_either(&self) -> Address<A, Either> {
         Address {
+            sink: self.sink.clone(),
             ref_counter: self.ref_counter.clone().into_either(),
-            sender: self.sender.clone(),
         }
     }
 
@@ -152,8 +153,8 @@ impl<A, Rc: RefCounter> Address<A, Rc> {
         if self.is_connected() {
             let (envelope, rx) = ReturningEnvelope::<A, M, <A as Handler<M>>::Return>::new(message);
             let tx = self
-                .sender
-                .clone()
+                .sink
+                .sender()
                 .into_send_async(AddressMessage::Message(Box::new(envelope)));
 
             SendFuture::sending_named(tx, rx)
@@ -196,17 +197,43 @@ impl<A, Rc: RefCounter> Address<A, Rc> {
         }
     }
 
-    /// Converts this address into a [futures `Sink`](https://docs.rs/futures/0.3/futures/io/struct.Sink.html).
-    pub fn into_sink(self) -> AddressSink<A, Rc> {
-        AddressSink {
-            sink: self.sender.clone().into_sink(),
-            ref_counter: self.ref_counter.clone(),
-        }
-    }
-
     /// Waits until this address becomes disconnected.
     pub fn join(&self) -> impl Future<Output = ()> + Send + Unpin {
         self.ref_counter.disconnect_notice()
+    }
+}
+
+impl<A, M, Rc> Sink<M> for Address<A, Rc>
+where
+    A: Handler<M>,
+    M: Send + 'static,
+    Rc: RefCounter,
+{
+    type Error = Disconnected;
+
+    fn poll_ready(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Pin::new(&mut self.sink)
+            .poll_ready(cx)
+            .map_err(|_| Disconnected)
+    }
+
+    fn start_send(mut self: Pin<&mut Self>, item: M) -> Result<(), Self::Error> {
+        let item = AddressMessage::Message(Box::new(NonReturningEnvelope::new(item)));
+        Pin::new(&mut self.sink)
+            .start_send(item)
+            .map_err(|_| Disconnected)
+    }
+
+    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Pin::new(&mut self.sink)
+            .poll_flush(cx)
+            .map_err(|_| Disconnected)
+    }
+
+    fn poll_close(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Pin::new(&mut self.sink)
+            .poll_close(cx)
+            .map_err(|_| Disconnected)
     }
 }
 
@@ -214,7 +241,7 @@ impl<A, Rc: RefCounter> Address<A, Rc> {
 impl<A, Rc: RefCounter> Clone for Address<A, Rc> {
     fn clone(&self) -> Self {
         Address {
-            sender: self.sender.clone(),
+            sink: self.sink.clone(),
             ref_counter: self.ref_counter.clone(),
         }
     }
@@ -226,7 +253,7 @@ impl<A, Rc: RefCounter> Drop for Address<A, Rc> {
         // We should notify the ActorManager that there are no more strong Addresses and the actor
         // should be stopped.
         if self.ref_counter.is_last_strong() {
-            let _ = self.sender.send(AddressMessage::LastAddress);
+            let _ = self.sink.sender().send(AddressMessage::LastAddress);
         }
     }
 }
